+++ im/benchmark_sender.go
@@ -5,27 +5,21 @@ import "net"
 import "log"
 import "runtime"
 import "flag"
+import "time"
 import "math/rand"
-import "net/http"
-import "encoding/base64"
-import "crypto/md5"
-import "strings"
-import "encoding/json"
-import "github.com/bitly/go-simplejson"
-import "io/ioutil"
+import "github.com/gomodule/redigo/redis"
+
 
 var first int64
 var last int64
 var host string
 var port int
+var redis_pool *redis.Pool
+var seededRand *rand.Rand
 
-
-
-const APP_ID = 7
-const APP_KEY = "sVDIlIiDUm7tWPYWhi6kfNbrqui3ez44"
-const APP_SECRET = "0WiCxAU1jh76SbgaaFC7qIaBPm2zkyM1"
-const URL = "http://192.168.33.10:5000"
-
+const redis_address = "127.0.0.1:6379"
+const redis_password = ""
+const redis_db = 0
 
 func init() {
 	flag.Int64Var(&first, "first", 0, "first uid")
@@ -35,43 +29,17 @@ func init() {
 	flag.IntVar(&port, "port", 23000, "port")
 }
 
-func login(uid int64) string {
-	url := URL + "/auth/grant"
-	secret := fmt.Sprintf("%x", md5.Sum([]byte(APP_SECRET)))
-	s := fmt.Sprintf("%d:%s", APP_ID, secret)
-	basic := base64.StdEncoding.EncodeToString([]byte(s))
-
-	v := make(map[string]interface{})
-	v["uid"] = uid
-
-	body, _ := json.Marshal(v)
-
-	client := &http.Client{}
-	req, _ := http.NewRequest("POST", url, strings.NewReader(string(body)))
-	req.Header.Set("Authorization", "Basic " + basic)
-	req.Header.Set("Content-Type", "application/json; charset=UTF-8")
-
-	res, err := client.Do(req)
-	if err != nil {
-		return ""
-	}
-	defer res.Body.Close()
-	
-	b, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		return ""
-	}
-	obj, err := simplejson.NewJson(b)
-	token, _ := obj.Get("data").Get("token").String()
-	return token
-}
 
 
 func send(uid int64) {
 	ip := net.ParseIP(host)
 	addr := net.TCPAddr{ip, port, ""}
 
-	token := login(uid)
+	token, err := login(uid)
+	if err != nil {
+		log.Println("login error")
+		return
+	}
 	
 	conn, err := net.DialTCP("tcp4", nil, &addr)
 	if err != nil {
@@ -111,5 +79,8 @@ func main() {
 		return
 	}
 	log.SetFlags(log.Lshortfile | log.LstdFlags)
+
+	seededRand = rand.New(rand.NewSource(time.Now().UnixNano()))
+	redis_pool = NewRedisPool(redis_address, redis_password, redis_db)
 	send(1)
 }
